requirements.txt
backcall==0.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
decorator==5.1.1
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
ipython==7.34.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyroute2==0.7.3
python-dotenv==0.21.0
requests==2.28.1
scapy==2.5.0rc1
traitlets==5.4.0
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.8.1
